+++ cli/src/definitions.ts
@@ -88,9 +88,22 @@ export interface AndroidConfig extends PlatformConfig {
   readonly resDir: string;
   readonly resDirAbs: string;
   readonly buildOutputDir: string;
+  /**
+   * @deprecated Will be removed in Cap. 5 as the `--flavor` option modifies this value.
+   */
   readonly buildOutputDirAbs: string;
+  /**
+   * @deprecated Will be removed in Cap. 5 as the `--flavor` option modifies this value.
+   */
   readonly apkName: string;
   readonly flavor: string;
+  readonly buildOptions: {
+    keystorePath?: string;
+    keystorePassword?: string;
+    keystoreAlias?: string;
+    keystoreAliasPassword?: string;
+    releaseType?: 'AAB' | 'APK';
+  };
 }
 
 export interface IOSConfig extends PlatformConfig {